requirements.txt
asgiref
certifi
charset-normalizer
Django
django-appconf
django-rest-framework
django-user-accounts
djangorestframework
idna
pip
pytz
requests
setuptools
sqlparse
urllib3
wheel
psycopg2-binary
dj-database-url
django-redis